Tests Nedelec Field Transfer from MFEM to PUMI.
Tests Nedelec Field Transfer for arbitrarily higher-order Nedelec
shape functions. A user defined function is used to project a
vector field onto the MFEM gridfunction. The vector field on MFEM
mesh is then projected onto the PUMI mesh by computing and storing
dofs on a PUMI nedelec field.
Both MFEM and PUMI fields are then evaluated at a xi-coordinate
to verify they both match exactly. Note that a xi-coordinate in
MFEM mesh might not be the same xi-coordinte in PUMI mesh,
because of different tet orientations which result due to the
reorientTetMesh() used in MFEM.
These orientations are taken care of when transferring the Nedelec
field and when evaluating the field at a certain xi-coordinate.